Fixed an issue that would make the global scene not to initialize in preview mode
